+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Next release
 
 HOOMD-blue 7.0 changes the definition of the rotation trial move size `a`. You will need to
 retune any hard-coded or saved values. Here are the distributions of the trial move rotation
-angles in HOOMD-blue 2.0--6.x and in 7.0+ for comparison:
+angles in HOOMD-blue 2.0–6.x and in 7.0+ for comparison:
 
 .. image:: https://github.com/user-attachments/assets/4a713aa9-2ec4-40dc-8363-aaf93e7eca2b