+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Next release
 
 *Added*
 
+* ``hoomd.hpmc.integrate.HPMCIntegrator.translate_move_dimensions``: Sets the dimensionality of HPMC translation moves, 
+  enabling interfacial confinement (xy) in 3D simulations.
+
 *Changed*
 
 * Use tabs in binary installation documentation (#2246).